Split tests into unit tests and integration tests
Ref. #252
