+++ tests/const.py
@@ -3,6 +3,8 @@ from pathlib import Path
 TEST_DIR = Path("tests")
 DATA_DIR = TEST_DIR / "data"
 CAPTURED_DIR = DATA_DIR / "captured"
+FIXTURE_DIR = DATA_DIR / "fixtures"
+IDE_DIR = TEST_DIR / "ide"
 SPEC_DIR = TEST_DIR / "specs"
 GENERATED_DIR = TEST_DIR / "spark" / "generated"
